+++ contracts/compliance/modular/modules/SpenderVerificationModule.sol
@@ -1,0 +1,158 @@
+// SPDX-License-Identifier: GPL-3.0
+//
+//                                             :+#####%%%%%%%%%%%%%%+
+//                                         .-*@@@%+.:+%@@@@@%%#***%@@%=
+//                                     :=*%@@@#=.      :#@@%       *@@@%=
+//                       .-+*%@%*-.:+%@@@@@@+.     -*+:  .=#.       :%@@@%-
+//                   :=*@@@@%%@@@@@@@@@%@@@-   .=#@@@%@%=             =@@@@#.
+//             -=+#%@@%#*=:.  :%@@@@%.   -*@@#*@@@@@@@#=:-              *@@@@+
+//            =@@%=:.     :=:   *@@@@@%#-   =%*%@@@@#+-.        =+       :%@@@%-
+//           -@@%.     .+@@@     =+=-.         @@#-           +@@@%-       =@@@@%:
+//          :@@@.    .+@@#%:                   :    .=*=-::.-%@@@+*@@=       +@@@@#.
+//          %@@:    +@%%*                         =%@@@@@@@@@@@#.  .*@%-       +@@@@*.
+//         #@@=                                .+@@@@%:=*@@@@@-      :%@%:      .*@@@@+
+//        *@@*                                +@@@#-@@%-:%@@*          +@@#.      :%@@@@-
+//       -@@%           .:-=++*##%%%@@@@@@@@@@@@*. :@+.@@@%:            .#@@+       =@@@@#:
+//      .@@@*-+*#%%%@@@@@@@@@@@@@@@@%%#**@@%@@@.   *@=*@@#                :#@%=      .#@@@@#-
+//      -%@@@@@@@@@@@@@@@*+==-:-@@@=    *@# .#@*-=*@@@@%=                 -%@@@*       =@@@@@%-
+//         -+%@@@#.   %@%%=   -@@:+@: -@@*    *@@*-::                   -%@@%=.         .*@@@@@#
+//            *@@@*  +@* *@@##@@-  #@*@@+    -@@=          .         :+@@@#:           .-+@@@%+-
+//             +@@@%*@@:..=@@@@*   .@@@*   .#@#.       .=+-       .=%@@@*.         :+#@@@@*=:
+//              =@@@@%@@@@@@@@@@@@@@@@@@@@@@%-      :+#*.       :*@@@%=.       .=#@@@@%+:
+//               .%@@=                 .....    .=#@@+.       .#@@@*:       -*%@@@@%+.
+//                 +@@#+===---:::...         .=%@@*-         +@@@+.      -*@@@@@%+.
+//                  -@@@@@@@@@@@@@@@@@@@@@@%@@@@=          -@@@+      -#@@@@@#=.
+//                    ..:::---===+++***###%%%@@@#-       .#@@+     -*@@@@@#=.
+//                                           @@@@@@+.   +@@*.   .+@@@@@%=.
+//                                          -@@@@@=   =@@%:   -#@@@@%+.
+//                                          +@@@@@. =@@@=  .+@@@@@*:
+//                                          #@@@@#:%@@#. :*@@@@#-
+//                                          @@@@@%@@@= :#@@@@+.
+//                                         :@@@@@@@#.:#@@@%-
+//                                         +@@@@@@-.*@@@*:
+//                                         #@@@@#.=@@@+.
+//                                         @@@@+-%@%=
+//                                        :@@@#%@%=
+//                                        +@@@@%-
+//                                        :#%%=
+//
+/**
+ *     NOTICE
+ *
+ *     The T-REX software is licensed under a proprietary license or the GPL v.3.
+ *     If you choose to receive it under the GPL v.3 license, the following applies:
+ *     T-REX is a suite of smart contracts implementing the ERC-3643 standard and
+ *     developed by Tokeny to manage and transfer financial assets on EVM blockchains
+ *
+ *     Copyright (C) 2025, Tokeny sàrl.
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+pragma solidity 0.8.30;
+
+import { IERC3643 } from "../../../ERC-3643/IERC3643.sol";
+import { IERC3643IdentityRegistry } from "../../../ERC-3643/IERC3643IdentityRegistry.sol";
+import { ErrorsLib } from "../../../libraries/ErrorsLib.sol";
+import { ModuleCapabilitiesLib } from "../../../libraries/ModuleCapabilitiesLib.sol";
+import {
+    AccessManagedOwnableBase,
+    AccessManagedOwnableUpgradeable
+} from "../../../utils/AccessManagedOwnableUpgradeable.sol";
+import { IModularCompliance } from "../IModularCompliance.sol";
+import { AbstractModuleUpgradeable } from "./AbstractModuleUpgradeable.sol";
+import { IModule } from "./IModule.sol";
+
+/// @title SpenderVerificationModule
+/// @dev Requires the spender of a `transferFrom` to be verified in the token's registry.
+/// Stateless: the registry is resolved through the compliance on every call.
+contract SpenderVerificationModule is AbstractModuleUpgradeable, AccessManagedOwnableUpgradeable {
+
+    constructor() {
+        _disableInitializers();
+    }
+
+    /// @dev Initializes the module behind its proxy.
+    /// @param accessManagerAddress authority gating the implementation upgrade
+    function initialize(address accessManagerAddress) external initializer {
+        require(accessManagerAddress != address(0), ErrorsLib.ZeroAddress());
+
+        __AbstractModule_init();
+        __AccessManaged_init(accessManagerAddress);
+    }
+
+    /// @inheritdoc IModule
+    /// @return true if the spender is verified in the registry of the bound token
+    function moduleCheckSpender(address _spender, address, address, uint256, address _compliance)
+        external
+        view
+        override
+        returns (bool)
+    {
+        return _identityRegistry(_compliance).isVerified(_spender);
+    }
+
+    /// @inheritdoc IModule
+    /// @return the bitmask of the dispatch points this module implements
+    function moduleCapabilities() external pure returns (uint256) {
+        return ModuleCapabilitiesLib.CHECK_SPENDER;
+    }
+
+    /// @inheritdoc IModule
+    /// @dev A compliance with no token gives the module no registry to read.
+    /// @param _compliance address of the compliance contract
+    /// @return true if a token is bound to `_compliance`
+    function canComplianceBind(address _compliance) external view returns (bool) {
+        return IModularCompliance(_compliance).getTokenBound() != address(0);
+    }
+
+    /// @inheritdoc IModule
+    /// @dev False so the bind check above runs: the compliance short-circuits
+    /// `isPlugAndPlay() || canComplianceBind(...)`.
+    /// @return always false
+    function isPlugAndPlay() external pure returns (bool) {
+        return false;
+    }
+
+    /// @inheritdoc IModule
+    /// @return _name the name of the module
+    function name() external pure returns (string memory _name) {
+        return "SpenderVerificationModule";
+    }
+
+    /// @inheritdoc AccessManagedOwnableBase
+    /// @param interfaceId the interface identifier, as specified in ERC-165
+    /// @return true if the contract implements `interfaceId`
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(AbstractModuleUpgradeable, AccessManagedOwnableBase)
+        returns (bool)
+    {
+        return AbstractModuleUpgradeable.supportsInterface(interfaceId)
+            || AccessManagedOwnableBase.supportsInterface(interfaceId);
+    }
+
+    /// @dev Resolves the registry of the token bound to a compliance.
+    /// @param _compliance address of the compliance contract
+    /// @return the identity registry of the token bound to `_compliance`
+    function _identityRegistry(address _compliance) private view returns (IERC3643IdentityRegistry) {
+        return IERC3643(IModularCompliance(_compliance).getTokenBound()).identityRegistry();
+    }
+
+    /// @dev Gated through the shared authority.
+    function _authorizeUpgrade(address) internal override restricted { }
+
+}